requirements.txt
# Core
streamlit>=1.31.0
langchain>=0.1.0
langchain-groq>=0.0.1
langgraph>=0.0.20
pydantic>=2.0.0

# APIs
tavily-python>=0.3.0
arxiv>=2.1.0
wikipedia>=1.4.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
